Renaming package: solver.backend.strategy to solver.strategy.
