+++ src/checkers/inference/solver/SolverEngine.java
@@ -17,8 +17,8 @@ import checkers.inference.model.Slot;
 import checkers.inference.model.VariableSlot;
 import checkers.inference.solver.backend.DefaultSolverFactory;
 import checkers.inference.solver.backend.SolverFactory;
-import checkers.inference.solver.backend.strategy.SolveStrategy;
-import checkers.inference.solver.backend.strategy.StrategyReflectiveFactory;
+import checkers.inference.solver.strategy.SolveStrategy;
+import checkers.inference.solver.strategy.StrategyReflectiveFactory;
 import checkers.inference.solver.util.Constants.SlotType;
 import checkers.inference.solver.util.PrintUtils;
 import checkers.inference.solver.util.SolverOptions;
