+++ src/checkers/inference/solver/strategy/AbstractSolveStrategy.java
@@ -1,4 +1,4 @@
-package checkers.inference.solver.backend.strategy;
+package checkers.inference.solver.strategy;
 
 import checkers.inference.solver.backend.SolverFactory;
