+++ src/checkers/inference/solver/strategy/PlainSolveStrategy.java
@@ -1,4 +1,4 @@
-package checkers.inference.solver.backend.strategy;
+package checkers.inference.solver.strategy;
 
 import java.util.Collection;
 import java.util.Map;
