+++ src/dataflow/solvers/general/DataflowGraphSolveStrategy.java
@@ -23,12 +23,12 @@ import checkers.inference.model.Slot;
 import checkers.inference.solver.backend.FormatTranslator;
 import checkers.inference.solver.backend.SolverAdapter;
 import checkers.inference.solver.backend.SolverFactory;
-import checkers.inference.solver.backend.strategy.GraphSolveStrategy;
 import checkers.inference.solver.constraintgraph.ConstraintGraph;
 import checkers.inference.solver.constraintgraph.GraphBuilder;
 import checkers.inference.solver.constraintgraph.Vertex;
 import checkers.inference.solver.frontend.LatticeBuilder;
 import checkers.inference.solver.frontend.TwoQualifiersLattice;
+import checkers.inference.solver.strategy.GraphSolveStrategy;
 import checkers.inference.solver.util.PrintUtils;
 import checkers.inference.solver.util.SolverOptions;
 import checkers.inference.solver.util.StatisticRecorder;
